Allow autoallocate service port, use it under minikube
We might very well (especially in development) want to run 2 or more
releases side by side and still use the nodePort functionality. In
order to achieve this, set the port to null and only pass the relevant
nodePort parameter if specified.
Stop asking for that parameter in the wizard since it is now useless
information and remove the now defunct namespace question as well

Change-Id: I0bb0b6dc6ae6c7c1f2336bcc328bc764c7ce9183