Helm test: Use environment variables for service-checker
Instead of relying on DNS discovery when doing helm tests, rely on the
kubernetes injected environmental variables that provide useful
information to the pod, one of which is the namespaces services.
Construct the correct variable name and use that for the
service-checker's arguments. Note that the parentheses are there for
kubernetes to do the expasion

Bug: T199489
Change-Id: I1fbbd63e0a9fef674d52c85f2c23a27ac54e2b9d