mathoid: Switch liveness probe into tcpSocket
Experiments have shown that having a liveness probe that relies on an
HTTP GET ends up in prematurely killing the pod when under stress since
it fails to service the probe in the configured timeout. That should not
happen as the pod is still serving requests. A better approach is to
depool the pod from the service and allow it to recover gracefully and
serve the remaining requests.

Experimental number for mathoid stress testing testing for the /png
endpoint with a nominal formula

For ncpu = 1:

With httpGet liveness probe:

* throughput: ~20 requests/s (this is an invariant)
* sustainable concurrent requests: ~22
* mean time per request: ~1250 ms

before the pod is killed by the kubelet

With a tcpSocket liveness probe on the other hand:

* throughput: ~20 requests/s (nominally the same)
* sustainable concurrent requests: > 250 (I stopped testing after that)
* mean time per request: ~12500 ms

For ncpu = 0

With httpGet liveness probe:

* throughput: ~20 requests/s (this is an invariant)
* sustainable concurrent requests: ~25
* mean time per request: ~1250 ms

before the pod is killed by the kubelet

With a tcpSocket liveness probe:

* throughput: ~20 requests/s (this is an invariant)
* sustainable concurrent requests: > 250 (I stopped testing after that)
* mean time per request: ~12100 ms

The numbers for concurrent requests in the 22 - 250 range are in the
expected range.

Of course the mean time per requests increases because more requests
just wait in the node queue waiting to be serviced. That's expected and
normal.

That also makes conceptually more sense as the liveness probe is about
the pod running and readiness probe is about the pod being able to
accept traffic.

A subsequent patch will also copy this logic in to the scaffold.

Change-Id: I909dc1b3cfe9be3f873ed5b97f3adedcee7dfc87