Add chart to pod labels
Add chart as well to the labels added to pods. This will allow to have
more granular matching for pods in the future as we plan to remove the
release label to allow for canaries. Matching also on the chart means
we can test new chart versions without messing up with the currently
deployed releases. Add that match to the corresponding Service stanza

Remove the erroneous TODO comment about doing canary at the deployment
level as we will be doing it at the service level

Update scaffolding as well

Change-Id: I5ecc5733e3185e78d0de829bd0cdd39a130cfc61