mathoid: Move config.yaml into a template
Move config.yaml into a template in order to be able to use stuff from
values.yaml. Actually use it to not instantiate a metrics stanza if
monitoring is disabled and to obtain the port the service should be
listening on from the chart. This is more DRY and future proof
One unfortunate caveat is the need to escape {{ and }} in go template.
Do this via the only sane helm way, namely {{ "{{" }}.
Amend the deployment checksumming code as well to allow the deployment
object to change when the config changes and hence a rolling deployment
to happen

Change-Id: Ibc81f5ad6a432474ce31bcddc61d9ebd43adddcf