Remove externalIP settings
Using externaIPs setting in services and using the same port as for the
nodePort as well as the service port (which is what happens in our
charts) causes a bind() race in kube-proxy where kube-proxy tries to
bind both to *:<nodePort> and <externalIP>:<port>. While those bind()
are immaterial for the actual implementation of the service as it is in
the end implemented by iptables, they are use by kube-proxy as a
failcheck mechanism in order to avoid misconfigurations sending traffic
to the wrong destinations. Stopping to use nodePort is currently a no-go
as we need kube-proxy to listen on the IP of the node for pybal to do
health-checks. So remove the externalIP setting for now and rely just on
nodePort. This on the plus side simplifies charts a bit, on the other
hand means we need to be a tad more careful when allocating ports in
order to avoid a conflict at the node level

Do this for all charts, bumping the chart versions and do so for
scaffolding as well

Change-Id: Ic9eb8858378952d8d397c35bda51e2d48803a731