Remove chartid from deployments/services
Adding the chartid to deployments/services unfortunately caused a side
effects of not allowing upgrades to newer chart version as the
deployment Selector field in apps/v1 API version became immutable. While
we could just remove that one instance, having it for the rest of the
fields would create confusion, so remove it fully. Do so for all charts
as well as scaffolding. Bump chart version to reflect that.

This partially reverts 49b53bd02

Change-Id: I399372264973c3d2b9aeed3b35594085cd7a1c74