Fix iteration of secret values in scaffolding
The previous code would use the dot (.) to generate the releasename and
in all cases this worked well, except in the generation of secrets.

The dot is set to the value of a key/value set when executed in the
scope of a range statement in gotemplate, which is what is happening in
deployments resources in this case. Set a temporary variable
representing the top scope and use it instead of the dot.

This affects blubberoid, citoid, mathoid, zotero as well as the
scaffolding.

That being said, none of the above services actually use secrets so
instead of maintaining that code, remove it from above said charts
While at it, also removing the public env vars iteration from
blubberoid, citoid, mathoid as they don't use it either. Amend
values.yaml for these charts as well

Change-Id: I5a1a6a0ff7b3b957341c3b56cc9715674972c486