Switch mathoid_requests_duration to a histogram
Switch from using a summary to using a histogram, in order to be able to
have aggregate quantiles from all instances of each service.
While at it, rename the metric to:

service_runner_request_duration_seconds

to have a more generic name that will be used by all service-runner
based services. Adopt that in citoid, eventgate-analytics and
scaffolding, alongside some sensible defaults

Fix the buckets and metrics names in some cases.
Fix the regex to be more explicit

Change-Id: I6ccb3aa1f912c87575b916fe0d5a72e6e8807293