blubberoid: Add policy file
Add a policy file to the blubberoid service. This uses the go-playground
validator to ensure that "base" images used for the production and test
variants come from a registry that is defined in
.Values.policy.docker_registry.

By default, the policy file is not enabled to allow for ease of local
development of the blubberoid chart.

The policyfile included in this chart makes use of the "startwith"
validator in go-playground/validator 9.29.0+ in blubber 0.8.0.

Bug: T215319
Change-Id: I7874a164b665240403749ce7ca2c04ae41197ef5