Give scaffold template configuration options for dev purposes
Devs using helm charts may need to be able to mount shared volumes,
expose additional ports, or run containers with different commands.
Ammended the templates and values file in the _scaffold directory
to allow this.

Changed the name of the configmap so it won't collide with
other configmaps if there are multiple subcharts under one release.

Bug: T226660
Change-Id: I622daffc90eef1ea650a60145bd5a7d1acb81ffe