helmfile,deploy: coredns needs api host and port
coredns needs to have KUBERNETES_SERVICE_HOST
and KUBERNETES_SERVICE_PORT environment variables
otherwise it would try to use the default kubernetes ip
and port.

In our current api-server setup, api certificate is generated by Puppet,
and puppet generated certificates lacks a SAN field including the ip of the api.
Coredns will refuse to validate the certificate if the SAN field is not present,
hence this change will instruct coredns to use the passed host instead of the IP
for validation.

Bug: T226516
Change-Id: I83e45dab2eff33d366abf68747f6f134e58869a6