Redeploy on stream-config changes
Instead of fetching stream-config.yaml as a file, invoke it as a
template, which is what it actually is, and run the hashing function on
the output. This correctly mutates on changes of the stream_config
values, mutating the deployment object as well and triggering a
redeployment.

Bump chart version and publish it to reflect that

Bug: T228700
Change-Id: I16d8431b3eac8a8b01450c6c94cb6f57cb10db9d