Add resources stanza to prometheus-metrics-exporter
Add a resources stanza based on the last 30days of data we have via
grafana. For limits add 3x the max() of either cpu or memory in order to
be both future safe and spike safe. For requests add the avg() for both
with some minor room to spare. The numbers are calculated across all of
the pods, they are not mathoid specific, so this is to be added to all
charts, mathoid is just the PoC

While at it, add a livenessProbe for the metrics exporter and the
informational ports stanza

Bug: T228837
Change-Id: I66a1f73b1526c19317bb7058b16d655839dc983a