mathoid: Take tiller into account as well
Since the LimitRange objects is per namespace, it will also encompass
tiller. While it does not consume many resources, it should be taken
into account, especially in the memory case.

While at it, lower the maximum pod size to be the app + the metrics
exporter, and lower the maximum amounts of pods as well, allowing up to
2 maximum sized pods in staging and up to 30 in production

Change-Id: Ib8f90cbd6a63514fd8ea98ca02ff035d1affa770