codfw: Bump all LimitRanges and ResourceQuotas
Instead of trying to figure out the best value per project for
LimitRanges and ResourceQuotas, do the following:

* set sensible minimums based on tiller and prometheus-metrics-exporter
container, which aren't going to be having requests/limits set anyway.
* Obtain the maximum from all present charts for requests/limits and set
double that as maximum. Use that value as is for pod maximum (the
assumption here is 1 container per pod if someone wants to push it that
hard).
* For ResourceQuotas, obtain that maximum and multiply it by 30, which
is the current maximum amount of replicas.

Bug: T228837
Change-Id: Ide6061d6be2e1b1196781a7562c743ea6db4be6d