Revert Add resources stanza to prometheus-metrics-exporter
There is no reason to have those at the end of the day. They create more
burden in the development environment than is worth and in production we
are going to rely on the LimitRanges default to provide those values
anyway. It was a nice PoC and useful, but now revert it

Bug: T228837
Change-Id: I2af45f75d8baf40dd30a2d13fdcf24db858a9dae