sessionstore: Add flag to disable TLS support
The raison d'etre for this the collection of profile information in
the staging environment ala `go tool pprof`.  Recent versions of the
tooling support specification of the ca, cert, and key when connecting
to a debug endpoint, but the version currently installed on deploy
hosts does not.

Having this option accidentally enabled in production would be Bad™;
We may want to consider reverting this after T229697 is closed.

Bug: T229697
Change-Id: I4d9fded6402bc8698178e98473d2f931b57d666e