blubberoid: Remove monitoring support
Currently blubberoid does not support neither prometheus, nor statsd.
Having support for it in the chart when the software does not yet
support either of the two gives wrong impressions and potentially causes
confusion. What's more, prometheus is instructed to scrape it in
production, which it can't do, reporting errors.

* Remove support for statsd prometheus exporter as well as all the
supporting things likes configmaps, networkpolicies etc.
* Remove the instrution in the chart for prometheus scraping

When support for monitoring is added, it should probably be done
natively in blubberoid via golang's excellent libraries. In that case,
this can be partially reverted, namely the scraping annocations and the
networkpolicy

Change-Id: I1df5dd7a33b6d531cc282c8065af0887b4b052dc