wikifeeds: Interpolate correctly the port value
.Values.main_app.port doesn't exist anymore since 00e3486a051 and
instead the ports array as well as appbase_url_port was introduced. Use
the later for the configuration as it is meant to be the "main" port of
wikifeeds

While at it, switch the monitoring default to false, in order to spare a
casual user of the app of having to deal with the repercussions of it,
e.g. extra memory usage, kubectl invocations requiring to pass container
names, etc

Release version 0.0.3 of the chart

Bug: T233076
Change-Id: If8bca759cdb57c0586df0b2685de3d55551457c4