Update scaffold template names to use chart name
Using the same name for templates between charts can cause
problems if charts are deployed in the same release, because
templates are global.

This change:

Updates the scaffold to use the chart name as a prefix for
templates instead of "wmf".

Updates create_new_service.sh to apply envsubst to all files using
templates in order to insert the chart name.

Adds a note about templates being global to the readme.

Change-Id: I8c9f4ed4341986187fbbe5c846156787afe0d279