RESTRouter: Skip resources on start-up and add nqo.wp.org
Both RESTBase and RESTRouter perform start-up checks against Cassandra
in order to ensure the environment is set up properly. However, in the
case of RESTRouter, that is not needed (as RESTBase does that already)
and expensive (as all of these requests need to be done over the
network). Therefore, tell RESTRouter not to perform them. Note that this
config option is currently a no-op as a new version of RESTRouter needs
to be available for that.

Also, add nqo.wp.org to the list of domains.

Note: there is no need to bump the chart version now because we have to
release a new version of the RESTRouter code base in order for these
changes to take effect, so we can bump it later when we change the image
tag.

Bug: T223953
Bug: T233833
Change-Id: Iaa4bd4da49d9a16b2a92cd7a055793866aaabf1f