Also check charts generated by helmfile
We don't want to just test fixtures, but also what we're going to apply
in our kubernetes clusters. Hence, run a separate task to check that
values presented in helmfiles generate a valid template as well. This
has the limitation that values coming from the private repository won't
be tested, but there is no way around that.

Change-Id: I50bdfcef6dc03b7283d916d2837da6d0288937e4