Create common template helpers directory
For some ancillary functions, like tls or debugging or even
monitoring, most charts share the same bits. It makes sense to group
them in a (filesystem-path versioned, for now) common_templates
directory and only include symlinks into the charts.

If we can be disciplined about not adding breaking changes to a
specific version of a common template, this gains us the following:

* DRY
* Ensuring bugfixes can reach the individual charts faster
* Consistency across charts

Bug: T235411
Change-Id: Ic34db64c2b16fa67000813b7a7c103f1b63c2b6e