cluster-helmfile: Add a simple sleep 1
Why:
It seems that if we do too many helmfile commands in very quick
succession, the kind that can only be done in a tight for loop,
occasionally tiller will barf and return an error. Add a simple sleep 1
statement to avoid this. This should be sufficient as this script is
not used that often to justify an actually good solution

What:
Add a simple sleep 1

Change-Id: Iac445f5097668f29b4ed331c29f3d8692b3c20a2