mathoid: Support canary functionality
Why:
We want to be able to have canary functionality in our releases, to
allow for small scale testing of new versions and configurations

What:
Allow a release to define a parent release via the canaryof value. When
that happens, don't instantiate a service for the release and label the
pods with the name of the specified parent release, allowing them to be
driven by the parent's service. Also alter the new network policy to
just match the canary pods in case it needs to be different than the
main one.
Bump the Chart version to reflect that

Change-Id: I66583d3dcd2e06273fe3a01a6dbd32a1210146b0