Remove all 1.7 netpol checks
Why:
It's no longer applicable. We are past way 1.7, our charts are only
compatible with 1.8+ and we aren't gonna be having an upgrade from 1.7
to 1.8+

What:
Remove the if clause from everywhere. It's no longer applicable. Also
fix a couple of indentation issues in a couple of charts. Those should
be a noop

Change-Id: I1fbf5fa61e9181ec14b4ff638bb305a51d54a55d