admin: Use a template to drop values symlinks
Why:
symlinks can be confusing, especially since they have the same names and
we want a clearer way to reason about this. In

What:
Move to a template in the helmfile to get the only interesting piece of
information for the top level helmfile, which is the names of the
environments (the namespaces essentially). Then drop the references to
envs.yaml from the top level helmfile.yaml and then the no longer
required symlinks. Amend the values file to account for the move

Change-Id: Ia24e9b9cd8fa718ac69a7a5429bbe08e2363b229