wikifeeds: Redefine CPU limits
Why:
During an outage we noticed that wikifeeds was severely CPU throttled.
Also it looks like it is throttled as well, ableit slightly in normal
operations

What:
Bump the request CPU by 25% to reach 1 millicore. That seems to match
better the average usage in production. Bump the limit to 3 cpus as this
seems to be the ceiling of throttled CPU during an outage
Bump the chart version to reflect the changes

Bug: T244535
Change-Id: I0bc616e082983d7df5ccf8938537d1b415fd9656