tls: Supply sane default values for resources
Why:
We found out that our envoy TLS proxy sidecars where in some cases
mildly throttled. We could increase the defaults per namespace and rely
on those, but that won't allow the charts to easily override those on a
need to basis.

What:
Add support for specifying resources stanza on the tls-proxy sidecar
After careful consideration and using https://w.wiki/K6N as a guide for
coming up with sane defaults for the sidecar resources, supply default
values as well.
Bump chart versions for those charts that use the _tls_helpers.tpl. The
ones that don't will have to be amended

Bug: T244843
Change-Id: I57a3b12058bfabb3bd497d3778faaae8ec5f3841