Prefix purely included template files with underscore
Why:
Templates that are only included generate almost null output when helm
template is run against them. However it's not entirely no output, there
is a YAML document containing a comment single line referring to them,
e.g.

  ---
  # Source: chromium-render/templates/config.yaml

This makes it more difficult to reason about helm template output,
especially for tools like kubeyaml
https://github.com/chuckha/kubeyaml

What:
Prefix the files containing those templates with an underscore as this
instructs helm to not output their result. However it is still being
included as is. This is a noop for all charts and has been tested for
all charts automatically. The only chart that needed a minor amend was
parsoid that did not address it's configuration properly

There will be no version bump for this reason

Change-Id: I78beae25487a6664b0bdf405ae9f6018dcbd633a