admin: Default to sensible values for deployUser, namespaceName
Why:
Having to create files of the format

  namespaceName: <service_name>
  deployUser: <service_name>

is not particularly useful, but rather mindless work. All files
currently have values that default to the name of the service anyway.
Making adding a new namespace less work is the goal here

What:
In case namespaceName or deployUser are not defined, default to the
name of the environment, that is the name of the service. We will be
killing those files in a followup commit

Change-Id: I2032c3509a8218e8f5f27391f4ab7da42f1a5dad