Make configuration of envoy a ConfigMap
There are various reasons not to keep using the more complex envoy image for
tls termination; most importantly, we want to implement a full services proxy
and that isn't viable without using a true templating language.

Instead of having a full installation of some templating system in the
container, we prefer to use helm's templating system to construct the
configmap.

This has two operational advantages:
- No need for derived docker images, or additional space usage
- The configuration is easily inspectable from the kubernetes api
itself.

Bug: T244843
Change-Id: Ie7f4c37508866de05e96eea104c8b8f6fa1e3454