Add local service proxy to the tls terminator v0.2
This adds the ability to add clusters and listeners to an envoy sidecar to allow
the application to use envoy for connection management, monitoring, tracing,
encryption and uniformity of timeouts.

This is done by using two data structures, one containing all the
upstream definitions (that could be common to all deployments), and
one including just what services should be proxied to from this
specific pod.

As long as we don't define those, nothing will change in the TLS
configuration of our services.

Bug: T244843
Change-Id: Ic2196161ccf976c925d11fc65d8f3c9de72b8c27