changeprop: enable nutcracker in production, reenable rate limiting
Before we can deploy changeprop to eqiad and codfw, we need to enable
Nutcracker. This configuration is currently working in staging and we
ened to enabled in prod to work in a similar manner.

Also reenable rate limiting for mw_purge. We previously disabled this
to debug redis issues that were caused by an unrelated
misconfiguration of nutcracker and this is safe to reenable
now. Changeprop has been repackaged to restore this behaviour also.

Bug: T213193
Change-Id: I8ffb55657a9da4d8396c0c2b2b4352e36b4b62c3