changeprop: Move prometheus query to use regex
Based on some testing, it seems that the prometheus graphite exporter
doesn't support matching using the default `glob` format for values
that aren't separated by dots. Given that we're matching beyond dots,
we need to use regex to get the slightly more complex matches we
require. Without this, the metrics exporter container will crash on
startup.

Bug: T213193
Change-Id: Ia9b1b777f2e9324e3a6df15780a0d64668973e27