changeprop: Make service features toggles rather than comments
This change is implemented in order to gradually enable features on a
per-environment basis and in a manner that won't require repeated
builds of the helm chart for changeprop.

Certain rules require other rules enabled - in order to avoid having
to manage dependencies, these rules are enabled in groups. This
behaviour means that enabling mw_purge (currently our default) also
enables null_edit and on_wikidata_description_change. This will mean
that this version of the chart will enable these rules when
deployed. Currently there is no need to deploy a new version to
production so we can let this change sit on staging for a period of
time, but if we want to make other changes to the helm chart for some
reason we will have to roll back this change if we are not comfortable
with it.

Bug: T248677
Change-Id: I09a50fb4ed19d290c40bb91feed7c6f76da21940