changeprop: Bump log level
changeprop's logging to stdout/stderr is quite terse at warn, and info
does not increase this massively but it will increase it enough to
give us more insight into how changeprop is behaving. If this volume
increases as we enable features we'll tune it downwards but for now
let's increase it.

Also fix the instructions for using the kakfa-dev subchart in
development.

Bug: T248677.
Change-Id: I07d198e8f9367290427d6ee6563b03c44808f813