Changeprop: add puppet CA cert to environment variables
Changeprop cannot talk to eventgate over HTTPS as it cannot verify the
puppet-signed server certificate. Node can load ca certificates via
the environment variable specified in this change. This is not an
ideal approach longterm but will allow us to at least test the
feasibility of this approach in the short term in staging.

Bug: T249633
Change-Id: I3122fbac968ec52c6d85dfa7b681a44cc1eb8829