eventgate, eventstreams: Log with namedlevels
Why:
Up to service-runner 2.7.7, logging to stdout would emit an integer log
level, something that while functional, would not be very nice to see in
logstash, nor would it convey any information (e.g. is "50" a warning or
an information message?). While this wasn't a problem when services
would directly log to logstash, in the logging pipeline that is becoming
a minor nuisance and would be nice to solve.

What:
Pass the named_levels: true parameter that service-runner 2.7.7 and
onwards supports

Bug: T239459
Change-Id: Iad0cb9cb1f8304a6083b9e12e0bd69f8728018f7