changeprop: add cpjobqueue configuration switching
rather than implement a subchart and the dependency management
entailed, or worse still a completely independent chart, let's simply
define cpjobqueue as a changeprop instance with a different
configuration. if cpjobqueue config isn't populated in values.yaml
then we assume we're a changeprop instance as normal. If it is then we
replace the changeprop configuration with the cpjobqueue
configuration. We'll do this *only* through values.yaml
per-cluster.

Bug: T220399
Change-Id: I9ee50295733a1395bf60ce646e0e414ff5eb373f