Add tool and configuration for generating beta configuration from kubernetes
This tool will use an existing checkout of our deployment-charts repo
to generate a helm template output based on defaults that fit our
existing beta configuration. When we update our production
configuration or the changeprop configuration file, we should use this
script to regenerate the changeprop beta config and commit it to this
repo.

Bug: T251176
Change-Id: Id22dffac26704005e8c3198fd7f3d3113f1c3066