changeprop: default all feature toggles to on
Up until now we have defaulted to disabling all rules in changeprop
and toggling them on as we roll out changeprop. Now that changeprop is
in a reasonably mature and stable state in production, these toggles
are now somewhat redundant. This change defaults to setting all
toggles on, and in future should they be needed we can toggle features
off explicitly.

Bug: T248677
Change-Id: I9d8e829b796d22be641b9dc692a140ae8f84cd7c