changeprop: simplify config writing. make beta config write puppet-friendly YAML.
This change removes the need for intermediary tempdirs for generated
files, we just generate the template we want and nothing else.

We also change how anchors are generated. Horizon will accept the YAML
we generate with this tool, but puppet itself when consuming it will
object to the presence of anchors as they will be interpreted by the
puppet YAML parser instead of being read as literal strings. As long
as this config is going into puppet metadata we can't coerce anchors
into strings or anything because that won't be valid YAML.

That said, it looks to me like we *can't* use Horizon for our
configuration in the first place as it'll clock in at well over the
current 64KB limit which might be an openstack-bound limit. Who
knows. If we don't use the anchors, our config is sufficiently small,
but it won't be loaded as unsafe YAML in puppet. If we do, it'll be
over the Horizon limit.

Bug: T251176
Change-Id: Iba99bf99d3cb6f8e480315b6b4772297834dc55e