tls_helper: Add a checksum for tls config (envoy)
Currently checksum/tls does not change on some config changes as their
values are not included into the checksum generated.

I choose to use the rendered config to generate one checksum
(tls-config) and the string representation of the tls.certs map (in
case something gets added in the future) for another (tls-certs).

Change-Id: I8c7c765f6eba0f72536bc8e908c3c9440378cdf2