eventgate: Port tls.upstream_timeout in values.yaml
Why:
_tls_helpers for eventgate diverged and need to be synced manually

What:
Set the same tls.upstream_timeout as the rest of the charts.
Interestingly, that value is set nowhere in our deploys and hence the
environment variable UPSTREAM_TIMEOUT ends up as an empty one.
Presumably that means the envoy will fallback to the 15s default[1] which
for this service is already quite a bit and hence the 180s won't hurt
either as it's way less aggressive

[1] https://www.envoyproxy.io/docs/envoy/latest/faq/configuration/timeouts#route-timeouts

Bug: TXXXXX
Change-Id: Id094999240f738c53c717df0764f07ec65738ab1