changeprop-jobqueue: enable all high-traffic jobs, increase replicas
We know that our kubernetes instances can talk to jobrunner without
issue. All of the high traffic jobs use jobrunner, so let's enable
them. Increase the number of replicas in prod to 12, which is the
number we have for changeprop. We might need to increase this number
as we enable further jobs as cpjobqueue handles a higher volume than
changeprop itself.

Bug: T220399
Change-Id: I98818b280cd326c07bdfacd0d5a75e426608f8be