changeprop-jobqueue: fix rendering of ignore topics list.
Previously we were rendering the values of the JSON dicts into the
exclude topics list, which totally messed things up. Instead, use the
keys of the dicts in the ignore list. This produces the desired
output.

Once we're sure that this is behaving as desired in staging we will
enable low traffic jobs in prod

Bug: T220399
Change-Id: I82920840e723ff77def9238858f8568c214fcc9a