mobileapps: Add a temporary non-TLS release
Why:
We want to separate the 2 migrations we currently have underway for
mobileapps. TLS and k8s. The reason for the separation is simply to make
the migration easier and with less thing to debug.

What:
Add a temporary release that does not have TLS enabled. Keep everything
the same as the normal release, except the name (even the port does not
need to differ as the nonTLS one does not listen on the standard port).
We will use first this release to switch to k8s and then after the TLS
migration is done, we will remove it

A different approach would be to have the production release listen on
both TLS and non-TLS ports, but that requires a chart bump. On the plus
side, this also allows us to verify our tooling still works fine with
multiple helm releases

Bug: T218733
Change-Id: Ibeb188e51f19e44bdd5fe63b8a35866bf1e5091a