mobileapps: add request template for load.php requests to config
In the course of switching over traffic from the old SCB deployment,
a bug was uncovered in which the service was making requests to the
public load.php endpoint. This succeeded on SCB but failed in the
Kubernetes deployment. The request is being updated in the service to
use a templated request definition; this updates the service config with
the needed template.

Bug: T258186
Change-Id: I7199771929ee8011c69d68627d249a9008cd749d