helmfile: refactoring blubberoid
This is just a basic proposal on how to refactor our helmfiles for
services. This changes a few things:

- Using environments and templates, we can respect the DRY principle
- Once we have a newer helm-diff, we can pass --kubeconfig as an arg
  and get rid of needing the environment variables to be defined
- we restructure the hierarchy of value files, with a base general one
  to apply to everything, a values / private files to apply to all
  releases in this helmfile, and per-env values to override the basics
  if needed.

The release workflow would change as follows:

Old:
$ cd staging/blubberoid
$ source .hfenv
$ helmfile -i sync
[rinse, repeat for codfw and eqiad]

New:
$ cd blubberoid
$ helmfile -i -e staging sync
$ helmfile -i -e eqiad sync
$ helmfile -i -e codfw sync

which is 5 less commands for every release.

This also means that changes will not be in any ways applied to the
wrong cluster.

Bug: T258572
Change-Id: I1f8e014ea03d00c8b6802f255871a98e6531553b