mobileapps: Lower replicas to 80 from 240
Why:
Clearly 240 was way too much. With about a day of data in prometheus,
the sum of max user and max system cpu usage is at 40 CPUs. Provide
double that to account for spikes. That is already 33% of the previous
value and a way better usage of resources. We 'll evaluate more later on
with more data

What:
Lower from 240 to 80

Bug: T218733
Change-Id: I7c724e693c2622d8e4618bf0439a3b7b3c7fca13