common_templates: Sort values for checksum/tls-certs
The map used to generate the checksum of tls certs is unordered, so we
end up with different hashes from time to time although nothing changed.

This used the sorted values of .Values.tls.certs as input for the hash
instead. While not perfect (as additional (unordered) maps in
.Values.tls.certs still lead to diverging hashes) it's better than what
we have now (and we currently don't have any other maps in
.Values.tls.certs).

Change-Id: I76882c12f143b2a449b7b115f6fd57b35e4e1942