eventgate - use /v1/_test/events route for readinessProbe
This route will produce the configured test_events as if they
were POSTed to /v1/events.  This is simpler than using the
included post-events script to exec and POST events.

Existing helmfiles will work with this new chart,
but the most recent eventgate-wikimedia version needs to be
deployed first.

For service that use local stream config (changing in T251935),
the chart will automatically configure test_events and
also a stream config entry to allow the configured test event
stream.

For service that use remote stream config (just eventgate-analytics-external ATM),
the test event stream MUST be configured in that remote stream config manually.
This has been done in https://gerrit.wikimedia.org/r/c/operations/mediawiki-config/+/618550

We should be able to merge and deploy this as is, and then change remaining services
to use remote stream config in a future patch.

Bump chart version to 0.2.7

Bug: T251935
Change-Id: I0d177eeebf9473b2db58a7274b7ff2208d296939