Support wikifeeds in api-gateway.
This does the following:
- Templates out the cluster type. For staging we actually point the
webserver cluster to mwdebug hosts, and having multiple hosts in
LOGICAL_DNS is not supported, we need STATIC_DNS
- Adds restbase_cluster upstream cluster via restbase.discovery
- Adds a /feed/... route with simple path regex rewriting. This
will not support feeds on commons/wikidata that do not have language
component in the path, but I believe we will need a major overhaul of
the path remapping anyway, so we can leave that for later. Trying to
regex all the special cases would be a nightmare.

Bug: T246265
Change-Id: I3a8a83a0f195fa21a2117822c6f0462f5814cc52