api-gateway: use x-client-ip instead of x-forwarded-for IP
Envoy doesn't always trust the client's IP address in the
x-forwarded-for header. Because we have the potential to see our k8s
node in the x-forwarded-for list, we might end up setting a ratelimit
for our kubernetes nodes, and that would be Real Bad. Use x-client-ip
which currently contains the correct client IP.

Bug: T246276
Change-Id: I66586813bedd82b605b0ab6d84eb0315e109857d