eventgate-logging-external - set schema_uri_query_param and stream_query_param
Setting these will instruct eventgate-wikimedia to
use the values of these query parameters as defaults
for the schema_uri_field and the stream_field if they
are not set in the event data.  This allows
the stream and schema for an event to be passed in
via the URL rather than needing to be in the event in the POST body.

Bug: T262087
Change-Id: I330761d2338a971ad2b930e46124c462026ef5ef