cpjobqueue: Increase cirrusSearchCheckerJob concurrency
Since transitioning from codfw back to eqiad these jobs have been
unable to keep up and had to be disabled to prevent a large backlog.
Various investigation has occured, but no concrete information is
available to determine the problem. With the jobs turned off, no new
data is generated to inform future efforts.

To get things going again increase concurrency of the jobs, with the
goal of bringing throughput back to prior levels.

Bug: T266762
Change-Id: I4845d38ffebcabe98fb4da0119dbea898e8981c1