coredns: Create a wmfcoredns copy in charts dir
This copies the "internal" chart wmfcoredns to where we have all other
charts to have them all in one place and to have them linted in all the
same way.

This does not change the current deployments in running clusters but is
rather meant to be installed during the cluster upgrades.

Changes to the chart itself:
* Remove the wmf prefix (it's obvious)
* Bump appVersion to the CoreDNS version used
* Use more clear variable names for kubernetes API
* Don't specify a clusterIP in chart (needs to be defined on a per
cluster basis anyways)
* Add fixture to have the chart linted with kubeyaml

Change-Id: Ibb671638408721681c3bd565127137f52e993f10