_tls_helpers: Add a default tls.image_version
Currently, if a chart does not define a value for tls.image_version
(which we should encurage, as we set the value globally via
/etc/helmfile-defaults), "helmfile lint" will fail.

With this change, the deployment can be linted without setting a value
for tls.image_version somewhere (that will get overridden anyways). As a
plus, the chart will work in minikube as well then.

Change-Id: I797821bc0893f3ddb9be96a0a0c30f284954c942