calico: Add GlobalNetworkPolicy support
Why:
We probably want to be able to set a GlobalNetworkPolicy to all namespaces.
It makes sense that we have a GlobalNetworkPolicy that is strict and
used as a fallback for all cases where we don't have a policy. This is
anyway the status quo with the current clusters.

What:
Add GlobalNetworkPolicy value and provide a "deny all except DNS" default for
all namespaces except kube-system (which contains calico, dns etc.).

Apps should be defining what they want in their NetworkPolicy resources.

Change-Id: I139f2d360f33a4cf84417239c9aeb7fee23f0e7f