Add a linkrecommendation-external release
Why:
Per T269581, an externally visible service would be beneficial. At the
same time, reusing the one that will be queried internally would be
problematic (details in the task). So, adding a dedicated release for it
makes sense.

What:
Add a new release that is identical to the main production one, aside
from being run with replicas = 1 and a different TLS port. The rest will
be inherited from the main service. The 1 replica should be enough to
satisfy 1 and 2 from the bullet point list in the task, but when 3 is
ready to take place we 'll need to reevaluate

Just so this is clear, this is an experiment.

Bug: T265603
Bug: T269581
Change-Id: I41a2ffe8cab7d5f5a2877ebb2f6f3f08465e724f