rdf-streaming-updater: switch to H/A session-cluster
the rdf-streaming-updater needs to communicate with both kafka and
swift which should be reflected in the network policy. The helm chart
has also been updater with config values as necessary.

Introduce H/A capabilities with a service account required to
manipulate configmaps on this namespace.

Switch to session-cluster and remove all settings related the
application from the chart.

Use the POD_IP to set the hostname the jobmanager uses for annoucing
itself to the taskmanagers, this is what is suggested from the
documentation.

Remove the graphite->prometheus sidecar container as flink has native
support for prometheus (see P16130 for an example output).

Add Values.main_app.config.additional_flink_settings to let
implementors customize flink (useful for testing locally).

Bug: T264006
Bug: T273098
Bug: T280166
Change-Id: Ic317f8a55654c3ab3aac84f7bb2487c3582f3c7a